Generalize inventory/item viewing/interacting.
Change InventoryCell into a general ItemCell function which simply
visualizes and allows mouse interaction with an Item. Alter
InventoryCell so that it no longer "uses" or "drops" the items, rather
it simply emits a signal telling you that the item was right or left
clicked. "Any computer science problem can be solved by adding a layer
of indirection, except the problem of having too many layers of
indirection."